5.1.10 Создаем класс PageController и в него выносим часть связанной логики.
